Change throughput computation and monitoring
Add `bool timeMultiplexing` to `getPairState` to determine how to
compute throughput.

Change `TCNDefaultBWLimit` from `int64_t` MB/s to `double` KB/s.

+++ src/server/services/optimizer/OptimizerService.cpp
@@ -100,7 +100,7 @@ void OptimizerService::runService()
     auto decreaseStep = config::ServerConfig::instance().get<int>("OptimizerDecreaseStep");
 
     auto qosInterval = config::ServerConfig::instance().get<int>("TimeMultiplexingQoSInterval");
-    auto defaultBwLimit = config::ServerConfig::instance().get<int>("TCNDefaultBwLimit");
+    auto defaultBwLimit = config::ServerConfig::instance().get<double>("TCNDefaultBwLimit");
 
     OptimizerNotifier optimizerCallbacks(
         config::ServerConfig::instance().get<bool>("MonitoringMessaging"),
